requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
dlib==19.24.4
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.13.4
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
Jinja2==3.1.3
MarkupSafe==2.1.5
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
platformdirs==4.2.0
proglog==0.1.10
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
scenedetect==0.6.3
soupsieve==2.5
sympy==1.12
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.2
transformers==4.40.0
typing_extensions==4.11.0
urllib3==2.2.1
videogrep==2.2.1
